+++ commands_extra/common/shell/check_github.sh
@@ -1,0 +1,29 @@
+#!/bin/bash
+################################################
+# Function : Check Github Site          
+# Desc     : 用于检查是否可以访问github网站脚本
+# Platform : ubuntu                                
+# Version  : 1.0                               
+# Date     : 2022-06-24 15:17:32                            
+# Author   : ncnynl                             
+# Contact  : 1043931@qq.com                              
+# URL: https://ncnynl.com                                   
+# QQ Qun: 创客智造B群:926779095                                 
+# QQ Qun: 创客智造C群:937347681                                  
+# QQ Qun: 创客智造D群:562093920                               
+################################################
+export TEXTDOMAINDIR=/usr/share/locale
+export TEXTDOMAIN=commands        
+echo "$(gettext "Check Github Site")"
+
+
+cd ~
+
+git clone https://github.com/ncnynl/test  rcm_check_github
+
+if [ -d ~/rcm_check_github ]; then
+    echo 1
+    rm -rf  ~/rcm_check_github
+else
+    echo 0
+fi