+++ test/proxy_github.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+
+git clone https://github.com/ncnynl/test  rcm_check_github